move jwt and alloy to pkg layer
